+++ bin/features/broker.js
@@ -5,8 +5,18 @@ const fs = require("fs");
 const path = require("path");
 const cds = require("@sap/cds");
 
-const { projectName, isNode, readYAML, generateHashBrokerPassword, adjustYAMLAllDocuments } = require("../common/util");
+const {
+  projectName,
+  isNode,
+  readYAML,
+  generateHashBrokerPassword,
+  adjustYAMLDocument,
+  adjustYAMLAllDocuments,
+} = require("../common/util");
 const YAML = require("yaml");
+const shelljs = require("shelljs");
+
+const config = require("../config.json");
 
 const BROKER_PATH = path.join(process.cwd(), "srv/broker.json");
 const CATALOG_PATH = path.join(process.cwd(), "srv/catalog.json");
@@ -83,6 +93,11 @@ function writeFile(path, content) {
 }
 
 module.exports = (options) => {
+  addXsuaaBroker();
+  if (!checkXsuaaBroker()) {
+    console.log(`Broker feature requires a project using xsuaa with service plan 'broker'`);
+    return;
+  }
   if (
     !(
       readYAML("mta.yaml")?.resources?.find(
@@ -166,3 +181,103 @@ module.exports = (options) => {
     );
   }
 };
+
+function addXsuaaBroker() {
+  const name = projectName();
+
+  // CF
+  adjustYAMLDocument("mta.yaml", (yaml) => {
+    for (const resource of yaml.get("resources").items) {
+      if (resource.getIn(["parameters", "service"]) === "xsuaa") {
+        resource.setIn(["parameters", "service-plan"], "application");
+      }
+    }
+    return yaml;
+  });
+
+  // Kyma
+  adjustYAMLDocument("chart/values.yaml", (yaml) => {
+    if (yaml.getIn(["xsuaa", "servicePlanName"])) {
+      yaml.setIn(["xsuaa", "servicePlanName"], "application");
+    }
+    return yaml;
+  });
+
+  // cds add
+  shelljs.exec(`cds add xsuaa ${config.options.cds}`);
+
+  // CF
+  adjustYAMLDocument("mta.yaml", (yaml) => {
+    for (const resource of yaml.get("resources").items) {
+      if (resource.getIn(["parameters", "service"]) === "xsuaa") {
+        resource.setIn(["parameters", "service-plan"], "broker");
+      }
+    }
+
+    // TODO: Remove (cap/issues/19545)
+    for (const module of yaml.get("modules").items) {
+      if (module.get("name") === `${name}-app-deployer`) {
+        const requires = module.get("requires");
+        if (requires && !requires.items.find((r) => r.get("name") === `${name}-auth`)) {
+          requires.items.push(yaml.createNode({ name: `${name}-auth` }));
+        }
+      } else if (module.get("name") === `${name}-destinations`) {
+        const requires = module.get("requires");
+        if (requires && !requires.items.find((r) => r.get("name") === `${name}-auth`)) {
+          requires.items.push(
+            yaml.createNode({
+              name: `${name}-auth`,
+              parameters: {
+                "service-key": {
+                  name: `${name}-auth-key`,
+                },
+              },
+            }),
+          );
+        }
+        const destinations = module.getIn(["parameters", "content", "instance", "destinations"]);
+        if (destinations && !destinations.items.find((d) => d.get("Name") === `${name}-auth`)) {
+          destinations.items.push(
+            yaml.createNode({
+              Name: `${name}-auth`,
+              Authentication: "OAuth2UserTokenExchange",
+              ServiceInstanceName: `${name}-auth`,
+              ServiceKeyName: `${name}-auth-key`,
+              "sap.cloud.service": `${name}.service`,
+            }),
+          );
+        }
+      }
+    }
+
+    return yaml;
+  });
+
+  // Kyma
+  adjustYAMLDocument("chart/values.yaml", (yaml) => {
+    if (yaml.getIn(["xsuaa", "servicePlanName"])) {
+      yaml.setIn(["xsuaa", "servicePlanName"], "broker");
+    }
+
+    // TODO: Remove (cap/issues/19545)
+    const bindings = yaml.getIn(["html5-apps-deployer", "bindings"]);
+    if (bindings && !bindings.get("xsuaa")) {
+      bindings.set(
+        "xsuaa",
+        yaml.createNode({
+          serviceInstanceName: "xsuaa",
+        }),
+      );
+    }
+
+    return yaml;
+  });
+}
+
+function checkXsuaaBroker() {
+  return (
+    readYAML("mta.yaml")?.resources?.find(
+      (r) => r?.parameters?.service === "xsuaa" && r?.parameters?.["service-plan"] === "broker",
+    ) || readYAML("chart/values.yaml")?.xsuaa?.servicePlanName === "broker"
+  );
+}