+++ scripts/deploy-java-cf.sh
@@ -1,5 +1,5 @@
 #!/bin/bash
 
 cd temp/afcjdk
-mbt build
-cf deploy mta_archives/afcjdk_1.0.0-SNAPSHOT.mtar
+npm install -s github:cap-js-community/sap-afc-sdk
+npx cds up --to cf