+++ scripts/deploy-node-cf.sh
@@ -1,5 +1,5 @@
 #!/bin/bash
 
 cd temp/afcsdk
-mbt build
-cf deploy mta_archives/afcsdk_1.0.0.mtar
+npm install -s github:cap-js-community/sap-afc-sdk
+npx cds up --to cf