+++ scripts/deploy-node-kyma.sh
@@ -1,7 +1,6 @@
 #!/bin/bash
 
 cd temp/afcsdk
+npm install -s github:cap-js-community/sap-afc-sdk
 docker system prune -a -f
-ctz containerize.yaml --push --logs
-helm upgrade --install afcsdk ./gen/chart -n afc-sdk-dev
-kubectl rollout restart deployment -n afc-sdk-dev
+npx cds up --to k8s -n afc-sdk-dev