+++ srv/scheduling/common/codelist.js
@@ -19,6 +19,7 @@ module.exports = {
   DataType: {
     string: "string",
     number: "number",
+    date: "date",
     datetime: "datetime",
     boolean: "boolean",
   },