+++ srv/scheduling/provider-service.js
@@ -7,6 +7,7 @@ const BaseApplicationService = require("../common/BaseApplicationService");
 const { JobStatus, MappingType, DataType } = require("./common/codelist");
 const JobSchedulingError = require("./common/JobSchedulingError");
 const { wildcard, toMap } = require("../../src/util/helper");
+const { approuterTenantUrl } = require("../../src/util/url");
 
 const isUUID = (input) =>
   input && /^[0-9a-f]{8}-[0-9a-f]{4}-[0-5][0-9a-f]{3}-[089ab][0-9a-f]{3}-[0-9a-f]{12}$/i.test(input);
@@ -15,11 +16,16 @@ module.exports = class SchedulingProviderService extends BaseApplicationService 
   async init() {
     const { Capabilities, JobDefinition, JobParameterDefinition, Job, JobParameter, JobResult, JobResultMessage } =
       this.entities;
-    const { JobDefinition: DBJobDefinition, Job: DBJob } = this.entities("scheduling");
+    const { JobDefinition: DBJobDefinition, Job: DBJob } = cds.entities("scheduling");
     const { notify } = this.operations;
 
-    this.on("READ", Capabilities, () => {
-      return cds.env.requires?.["sap-afc-sdk"]?.capabilities ?? {};
+    this.fillLink(Job, "Job", "manage");
+
+    this.on("READ", Capabilities, (req) => {
+      return {
+        ...(cds.env.requires?.["sap-afc-sdk"]?.capabilities ?? {}),
+        applicationUrl: cds.env.requires?.["sap-afc-sdk"]?.ui?.link ? approuterTenantUrl(req) : null,
+      };
     });
 
     this.before("READ", [JobParameterDefinition, JobParameter, JobResultMessage], (req) => {
@@ -269,6 +275,7 @@ module.exports = class SchedulingProviderService extends BaseApplicationService 
               }
               parameter.value = parsedValue;
               break;
+            case "date":
             case "datetime":
               parsedValue = new Date(parameter.value);
               if (isNaN(parsedValue.getTime())) {
@@ -281,6 +288,9 @@ module.exports = class SchedulingProviderService extends BaseApplicationService 
                 );
               }
               parameter.value = parsedValue.toISOString();
+              if (jobParameterDefinition.dataType_code === "date") {
+                parameter.value = parameter.value.split("T")[0];
+              }
               break;
             case "boolean":
               parsedValue = String(parameter.value);
@@ -414,7 +424,7 @@ module.exports = class SchedulingProviderService extends BaseApplicationService 
   }
 
   async createJob(req, job) {
-    const { Job: DBJob } = this.entities("scheduling");
+    const { Job: DBJob } = cds.entities("scheduling");
     for (const parameter of job.parameters) {
       if (parameter.value !== null) {
         parameter.value = String(parameter.value);
@@ -424,12 +434,12 @@ module.exports = class SchedulingProviderService extends BaseApplicationService 
   }
 
   async updateJob(req, job, data) {
-    const { Job: DBJob } = this.entities("scheduling");
+    const { Job: DBJob } = cds.entities("scheduling");
     await UPDATE.entity(DBJob).set(data).where({ ID: job.ID });
   }
 
   async downloadData(req, ID) {
-    const { JobResult: DBJobResult } = this.entities("scheduling");
+    const { JobResult: DBJobResult } = cds.entities("scheduling");
     const { data } = await SELECT.one.from(DBJobResult).columns("data").where({ ID });
     return {
       value: data,